pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ai-frontier"
version = "3.0.0"
description = "AI Frontier: local-first, self-improving transformer training system"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0.1",
    "psutil>=5.9.0",
    "transformers>=4.36.0",
    "sentencepiece>=0.1.99",
    "datasets>=2.16.0",
    "huggingface_hub>=0.20.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "tensorboard>=2.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[project.scripts]
ai-frontier = "scripts.launcher:main"

[tool.setuptools.packages.find]
include = ["frontier*", "training*", "evaluation*", "inference*", "dashboard*", "dataset_pipeline*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
